migrating to Eio library for IO/filesystem interaction
This commit replaces a bunch of code that used the Sys and Unix modules
with the Eio library. One of the benefits is that it makes forester
closer to being portable to non-unix systems.

I have added a new command-line option, --max-fibers=NUM, to control
how many fibers can be used for the LaTeX compilation pipeline. It can
be necessary to control this to avoid system errors involving too many
files open at a time.

I have also added the --ignore-tex-cache option.